requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=4.9.0

# Optional: For enhanced testing
pytest>=7.4.0
